# backend/services/contracts/internal-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Vocabulary Internal Service Contracts
  version: 1.1.0
  description: |
    Source of truth for `/internal/v1/*` contracts used between TypeScript owner-services,
    the public gateway, and Python capability APIs.
servers:
  - url: /internal/v1
paths:
...
  /nlp/extract-sentence:
    post:
      summary: Extract example sentence for a term.
      operationId: extractSentence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtractSentenceRequest'
      responses:
        '200':
          description: Example sentence payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractSentenceResult'
...
  /assignments/scan:
    post:
      summary: Synchronous assignment scan.
      operationId: scanAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentScanRequest'
      responses:
        '200':
          description: Scan result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentScanResult'
...
components:
...
  schemas:
...
    ExtractSentenceRequest:
      type: object
      required:
        - text
        - term
      properties:
        text:
          type: string
        term:
          type: string
    ExtractSentenceResult:
      type: object
      required:
        - sentence
      properties:
        sentence:
          type: string
...
    AssignmentScanRequest:
      type: object
      required:
        - content_completed
      properties:
        title:
          type: string
          default: ''
        content_original:
          type: string
          default: ''
        content_completed:
          type: string
    AssignmentMatch:
      type: object
      required:
        - entry_id
        - term
        - category
        - source
        - occurrences
      properties:
        entry_id:
          type: integer
        term:
          type: string
        category:
          type: string
        source:
          type: string
        occurrences:
          type: integer
    AssignmentMissingWord:
      type: object
      required:
        - term
        - occurrences
        - example_usage
      properties:
        term:
          type: string
        occurrences:
          type: integer
        example_usage:
          type: string
    AssignmentDiffChunk:
      type: object
      required:
        - operation
        - original_text
        - completed_text
      properties:
        operation:
          type: string
        original_text:
          type: string
        completed_text:
          type: string
    AssignmentScanResult:
      type: object
      required:
        - assignment_id
        - title
        - content_original
        - content_completed
        - word_count
        - known_token_count
        - unknown_token_count
        - lexicon_coverage_percent
        - assignment_status
        - message
        - duration_ms
        - matches
        - missing_words
        - diff_chunks
      properties:
        assignment_id:
          type:
            - integer
            - 'null'
        title:
          type: string
        content_original:
          type: string
        content_completed:
          type: string
        word_count:
          type: integer
        known_token_count:
          type: integer
        unknown_token_count:
          type: integer
        lexicon_coverage_percent:
          type: number
        assignment_status:
          type: string
        message:
          type: string
        duration_ms:
          type: number
        matches:
          type: array
          items:
            $ref: '#/components/schemas/AssignmentMatch'
        missing_words:
          type: array
          items:
            $ref: '#/components/schemas/AssignmentMissingWord'
        diff_chunks:
          type: array
          items:
            $ref: '#/components/schemas/AssignmentDiffChunk'
...